requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
attrs==24.2.0
    # via
    #   outcome
    #   trio
certifi==2024.8.30
    # via
    #   requests
    #   selenium
cffi==1.17.1
    # via trio
charset-normalizer==3.4.0
    # via requests
dnspython==2.7.0
    # via pymongo
fuzzywuzzy==0.18.0
    # via -r .\requirements.in
h11==0.14.0
    # via wsproto
idna==3.10
    # via
    #   requests
    #   trio
numpy==2.1.2
    # via pandas
outcome==1.3.0.post0
    # via trio
packaging==24.1
    # via webdriver-manager
pandas==2.2.3
    # via -r .\requirements.in
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via -r .\requirements.in
pydantic-core==2.23.4
    # via pydantic
pymongo==4.10.1
    # via -r .\requirements.in
pysocks==1.7.1
    # via urllib3
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via webdriver-manager
pytz==2024.2
    # via pandas
pywin32==308
    # via -r .\requirements.in
pyyaml==6.0.2
    # via -r .\requirements.in
requests==2.32.3
    # via
    #   -r .\requirements.in
    #   webdriver-manager
selenium==4.25.0
    # via -r .\requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.27.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
    #   selenium
tzdata==2024.2
    # via pandas
urllib3[socks]==2.2.3
    # via
    #   -r .\requirements.in
    #   requests
    #   selenium
webdriver-manager==4.0.2
    # via -r .\requirements.in
websocket-client==1.8.0
    # via selenium
wsproto==1.2.0
    # via trio-websocket
